group3_project/calendar/requirements.txt

annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.7.0
distro==1.9.0
Django==4.2.11
django-taggit==6.1.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
Markdown==3.7
openai==1.70.0
packaging==24.2
pluggy==1.5.0
pydantic==2.9.0
pydantic_core==2.23.2
pytest==8.3.4
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
setuptools==78.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
supervisor==4.2.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wheel==0.42.0
gunicorn==23.0.0
flake8==7.0.0
safety==3.3.1
